+++ applovin_max/android/src/main/java/com/applovin/applovin_max/AppLovinMAXAdView.java
@@ -23,19 +23,38 @@ import io.flutter.plugin.platform.PlatformView;
 public class AppLovinMAXAdView
         implements PlatformView
 {
-    private static final Map<String, AppLovinMAXAdViewWidget> widgetInstances          = new HashMap<>( 2 );
-    private static final Map<String, AppLovinMAXAdViewWidget> preloadedWidgetInstances = new HashMap<>( 2 );
+    private static final Map<Integer, AppLovinMAXAdViewWidget> widgetInstances          = new HashMap<>( 2 );
+    private static final Map<Integer, AppLovinMAXAdViewWidget> preloadedWidgetInstances = new HashMap<>( 2 );
 
     @Nullable
     private AppLovinMAXAdViewWidget widget;
+    private int                     adViewId;
 
     private final MethodChannel channel;
 
+    // Returns an MaxAdView to support Amazon integrations. This method returns the first instance
+    // that matches the Ad Unit ID, consistent with the behavior introduced when this feature was
+    // first implemented.
+    @Nullable
     public static MaxAdView getInstance(final String adUnitId)
     {
-        AppLovinMAXAdViewWidget widget = preloadedWidgetInstances.get( adUnitId );
-        if ( widget == null ) widget = widgetInstances.get( adUnitId );
-        return ( widget != null ) ? widget.getAdView() : null;
+        for ( Map.Entry<Integer, AppLovinMAXAdViewWidget> entry : preloadedWidgetInstances.entrySet() )
+        {
+            if ( entry.getValue().getAdUnitId().equals( adUnitId ) )
+            {
+                return entry.getValue().getAdView();
+            }
+        }
+
+        for ( Map.Entry<Integer, AppLovinMAXAdViewWidget> entry : widgetInstances.entrySet() )
+        {
+            if ( entry.getValue().getAdUnitId().equals( adUnitId ) )
+            {
+                return entry.getValue().getAdView();
+            }
+        }
+
+        return null;
     }
 
     public static void preloadWidgetAdView(final String adUnitId,
@@ -48,15 +67,8 @@ public class AppLovinMAXAdView
                                            final AppLovinSdk sdk,
                                            final Context context)
     {
-        AppLovinMAXAdViewWidget preloadedWidget = preloadedWidgetInstances.get( adUnitId );
-        if ( preloadedWidget != null )
-        {
-            result.error( AppLovinMAX.TAG, "Cannot preload more than once for a single Ad Unit ID.", null );
-            return;
-        }
-
-        preloadedWidget = new AppLovinMAXAdViewWidget( adUnitId, adFormat, true, sdk, context );
-        preloadedWidgetInstances.put( adUnitId, preloadedWidget );
+        AppLovinMAXAdViewWidget preloadedWidget = new AppLovinMAXAdViewWidget( adUnitId, adFormat, true, sdk, context );
+        preloadedWidgetInstances.put( preloadedWidget.hashCode(), preloadedWidget );
 
         preloadedWidget.setPlacement( placement );
         preloadedWidget.setCustomData( customData );
@@ -65,25 +77,25 @@ public class AppLovinMAXAdView
 
         preloadedWidget.loadAd();
 
-        result.success( null );
+        result.success( preloadedWidget.hashCode() );
     }
 
-    public static void destroyWidgetAdView(final String adUnitId, final Result result)
+    public static void destroyWidgetAdView(final int adViewId, final Result result)
     {
-        AppLovinMAXAdViewWidget preloadedWidget = preloadedWidgetInstances.get( adUnitId );
+        AppLovinMAXAdViewWidget preloadedWidget = preloadedWidgetInstances.get( adViewId );
         if ( preloadedWidget == null )
         {
-            result.error( AppLovinMAX.TAG, "No widget found to destroy", null );
+            result.error( AppLovinMAX.TAG, "No preloaded AdView found to destroy", null );
             return;
         }
 
         if ( preloadedWidget.hasContainerView() )
         {
-            result.error( AppLovinMAX.TAG, "Cannot destroy - currently in use", null );
+            result.error( AppLovinMAX.TAG, "Cannot destroy - the preloaded AdView is currently in use", null );
             return;
         }
 
-        preloadedWidgetInstances.remove( adUnitId );
+        preloadedWidgetInstances.remove( adViewId );
 
         preloadedWidget.detachAdView();
         preloadedWidget.destroy();
@@ -93,6 +105,7 @@ public class AppLovinMAXAdView
 
     public AppLovinMAXAdView(final int viewId,
                              final String adUnitId,
+                             final int adViewId,
                              final MaxAdFormat adFormat,
                              final boolean isAutoRefreshEnabled,
                              @Nullable final String placement,
@@ -108,12 +121,12 @@ public class AppLovinMAXAdView
         channel.setMethodCallHandler( (call, result) -> {
             if ( "startAutoRefresh".equals( call.method ) )
             {
-                widget.setAutoRefresh( true );
+                widget.setAutoRefreshEnabled( true );
                 result.success( null );
             }
             else if ( "stopAutoRefresh".equals( call.method ) )
             {
-                widget.setAutoRefresh( false );
+                widget.setAutoRefreshEnabled( false );
                 result.success( null );
             }
             else
@@ -122,27 +135,33 @@ public class AppLovinMAXAdView
             }
         } );
 
-        widget = preloadedWidgetInstances.get( adUnitId );
+        widget = preloadedWidgetInstances.get( adViewId );
         if ( widget != null )
         {
             // Attach the preloaded widget if possible, otherwise create a new one for the
             // same adUnitId
             if ( !widget.hasContainerView() )
             {
-                widget.setAutoRefresh( isAutoRefreshEnabled );
+                AppLovinMAX.d( "Mounting the preloaded AdView (" + adViewId + ") for Ad Unit ID " + adUnitId );
+
+                this.adViewId = adViewId;
+                widget.setAutoRefreshEnabled( isAutoRefreshEnabled );
                 widget.attachAdView( this );
                 return;
             }
         }
 
         widget = new AppLovinMAXAdViewWidget( adUnitId, adFormat, sdk, context );
-        widgetInstances.put( adUnitId, widget );
+        this.adViewId = widget.hashCode();
+        widgetInstances.put( this.adViewId, widget );
+
+        AppLovinMAX.d( "Mounting a new AdView (" + this.adViewId + ") for Ad Unit ID " + adUnitId );
 
         widget.setPlacement( placement );
         widget.setCustomData( customData );
         widget.setExtraParameters( extraParameters );
         widget.setLocalExtraParameters( localExtraParameters );
-        widget.setAutoRefresh( isAutoRefreshEnabled );
+        widget.setAutoRefreshEnabled( isAutoRefreshEnabled );
 
         widget.attachAdView( this );
         widget.loadAd();
@@ -170,15 +189,19 @@ public class AppLovinMAXAdView
         {
             widget.detachAdView();
 
-            AppLovinMAXAdViewWidget preloadedWidget = preloadedWidgetInstances.get( widget.getAdView().getAdUnitId() );
+            AppLovinMAXAdViewWidget preloadedWidget = preloadedWidgetInstances.get( adViewId );
 
             if ( widget == preloadedWidget )
             {
-                widget.setAutoRefresh( false );
+                AppLovinMAX.d( "Unmounting the preloaded AdView (" + adViewId + ") for Ad Unit ID " + widget.getAdUnitId() );
+
+                widget.setAutoRefreshEnabled( false );
             }
             else
             {
-                widgetInstances.remove( widget.getAdView().getAdUnitId() );
+                AppLovinMAX.d( "Unmounting the AdView (" + adViewId + ") to destroy for Ad Unit ID " + widget.getAdUnitId() );
+
+                widgetInstances.remove( adViewId );
                 widget.destroy();
             }
         }